Updating form functionality, wiring

+++ app/src/components/Motorcycle.vue
@@ -1,15 +1,24 @@
 <template>
   <div class="content">
-    <h2>{{ motorcycle.make }}</h2>
-    <p>{{ motorcycle.type }}</p>
-    <p>Model: {{ motorcycle.model }}</p>
-    <p>Year: {{ motorcycle.year }}</p>
-    <p>Color: {{ motorcycle.color }}</p>
-    <p v-if="motorcycle.available">Status: FOR SALE!</p>
-    <p v-else>Status: Contact us for availability</p>
-    <p>Delete:
-      <input @click="handleDelete()" type="checkbox" name="delete">
-    </p>
+    <article v-if="!editing">
+      <h2>{{ motorcycle.make }}</h2>
+      <p>{{ motorcycle.type }}</p>
+      <p>Model: {{ motorcycle.model }}</p>
+      <p>Year: {{ motorcycle.year }}</p>
+      <p>Color: {{ motorcycle.color }}</p>
+      <p v-if="motorcycle.available">Status: FOR SALE!</p>
+      <p v-else>Status: Contact us for availability</p>
+      <p>Delete:
+        <input @click="handleDelete()" type="checkbox" name="delete">
+      </p>
+    </article>
+    <MotorcycleForm
+    v-else
+    label="Update"
+    :neighborhood="neighborhood"
+    :on-edit="onUpdate"
+    />
+    <button @click="editing = !editing">{{ editing ? 'Cancel' : 'Edit Item' }}</button>
   </div>
 </template>
 
@@ -17,17 +26,13 @@
 
 export default {
 
-  props: {
-    motorcycle: Object,
-    onRemove: {
-      type: Function,
-      required: true
-    }
-  },
+  props: ['motorcycle', 'onRemove', 'onUpdate'],
 
   methods: {
     handleDelete() {
-      this.onRemove(this.motorcycle);
+      if(confirm(`Are you sure you want to remove this ${this.motorcycle.make} (${this.motorcycle.model})?`)) {
+        this.onRemove(this.motorcycle);
+      }
     }
   }
